requirements.txt
streamlit==1.40.2
langchain==0.3.9
langchain-openai==0.2.10
langchain-community==0.3.8
openai==1.55.3
duckduckgo-search==6.3.7
pyowm==3.3.0
yfinance==0.2.50
python-dotenv==1.0.1